+++ test/test_statement.cpp
@@ -14,9 +14,9 @@ using namespace thinsqlitepp;
 using namespace std;
 
 #if __cpp_lib_ranges >= 201911L
-static_assert(std::forward_iterator<row_iterator>);
+static_assert(std::input_iterator<row_iterator>);
 static_assert(std::ranges::random_access_range<row>);
-static_assert(std::ranges::forward_range<row_range>);
+static_assert(std::ranges::input_range<row_range>);
 #endif
 
 TEST_SUITE_BEGIN("statement");